backend/requirements.txt
flask
google-cloud-secret-manager
sqlalchemy
sqlalchemy-cockroachdb
psycopg2
python-dotenv
